requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3
pydantic==2.5.0
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1
transformers==4.35.2
torch>=2.2.0
accelerate==0.24.1
sentencepiece==0.1.99
httpx==0.25.2 